timer

+++ MultiEnemy.h
@@ -6,7 +6,7 @@
 #include "Score.h"
 
 // seconds
-const int MAX_DIFF_TIME = 1;
+const double MAX_DIFF_TIME = 1;
 
 class MultiEnemy {
 public:
